ReadMatlabDocstrings

Documentation for ReadMatlabDocstrings.

ReadMatlabDocstrings._getrawcomment

ReadMatlabDocstrings.getdoc

ReadMatlabDocstrings.lang_matlab

ReadMatlabDocstrings._getrawcomment — Method

_getrawcomment(s) get all strings before the line starts with "function" or "classdef"; the comment symbol such as "% " will be kept.

ReadMatlabDocstrings.getdoc — Method

getdoc(filepath::String; first_head=0) read the function descriptions and return a single string ready for Markdown.parse. Now it supports matlab's .m function only.

Example (matlab)

y = getdoc("..\..\src\statind.m") get the docstring (the comments (starts with % eachline) before "function statind(...)").

y = getdoc("..\..\src\statind.m", first_head=3) get the docstring and rearrange all the heading levels based on forcing the first heading level to be 3. That is, take the first heading to be # for example, # becomes ### and ## becomes ####

y = getdoc("..\..\src\statind.m", filename_header=3) get the docstring and insert a level-3 markdown header at the first line: "### statind".

ReadMatlabDocstrings.lang_matlab — Method

Given a vector of rawcomments, lang_matlab strips all comment symbols (i.e., %).